Fix math functions and clean up code
The math functions were rounding to the first decimal place every
calculation. This could lead to some incorrect outputs, so I reworked
the functions to only round when the solution button is pressed.

The solution and clear buttons were both setting variables to null
which cluttered the code. The solution button now calls the clear
button and tells it not to clear the screen, just the variables.